moving parity tests
